Move `LanguageRegistry` into `buffer`
