+++ zed/src/main.rs
@@ -32,7 +32,7 @@ fn main() {
 
     let themes = settings::ThemeRegistry::new(Assets, app.font_cache());
     let (settings_tx, settings) = settings::channel(&app.font_cache(), &themes).unwrap();
-    let languages = Arc::new(language::LanguageRegistry::new());
+    let languages = Arc::new(language::build_language_registry());
     languages.set_theme(&settings.borrow().theme.syntax);
 
     app.run(move |cx| {
